todo_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.1.0
drf-yasg==1.21.7
twilio==8.10.0
firebase-admin==6.2.0
django-filter==23.5
